requirements.txt
# enter requirements that aren't related to connecting to the messagebus here

stomp.py==4.1.20
PyYAML==3.12
pytz==2018.4
pyzmq==17.1.2
numpy==1.14.5
scipy==1.1.0
pandas==0.22.0
tabulate==0.8.2
pvlib==0.6.0
scikit-learn==0.20.1
influxdb